+++ tests/TestSupport/Components/Steps/InvalidStepComponent.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Spatie\LivewireWizard\Tests\TestSupport\Components\Steps;
+
+use Livewire\Component;
+
+// should extend StepComponent
+class InvalidStepComponent extends Component
+{
+}