+++ tests/WizardTest.php
@@ -8,6 +8,8 @@ use Spatie\LivewireWizard\Tests\TestSupport\Components\MyWizardComponent;
 use Spatie\LivewireWizard\Tests\TestSupport\Components\Steps\FirstStepComponent;
 use Spatie\LivewireWizard\Tests\TestSupport\Components\Steps\SecondStepComponent;
 use Spatie\LivewireWizard\Tests\TestSupport\Components\Steps\ThirdStepComponent;
+use Spatie\LivewireWizard\Tests\TestSupport\Components\WizardWithInvalidStepComponent;
+use function Spatie\Snapshots\assertMatchesHtmlSnapshot;
 
 beforeEach(function () {
     $this->wizard = Livewire::test(MyWizardComponent::class);
@@ -15,7 +17,7 @@ beforeEach(function () {
     $this->firstStep = Livewire::test(FirstStepComponent::class);
 });
 
-it('can render a wizard component', function () {
+$it = it('can render a wizard component', function () {
     $this->wizard->assertSuccessful();
 });
 
@@ -66,6 +68,10 @@ it('throws an exception when going to the next step on the last step', function 
         ->emitEvents()->in($wizard);
 })->throws(NoNextStep::class);
 
+it('will throw an exception if the wizard contains an invalid step', function () {
+    Livewire::test(WizardWithInvalidStepComponent::class);
+})->throws('did return an invalid step component');
+
 it('will save and restore state when switching steps', function () {
     $this->firstStep
         ->call('nextStep')
@@ -89,3 +95,50 @@ it('cannot set state if step does not exist', function() {
     $this->wizard
         ->call('setStepState', 'fake-step', []);
 })->throws(StepDoesNotExist::class);
+
+it('has a couple of handy methods to get state', function () {
+    $this->firstStep
+        ->call('nextStep')
+        ->emitEvents()->in($this->wizard);
+
+    $this->wizard->assertSee(['second step', 'counter: 0']);
+
+    Livewire::test(SecondStepComponent::class)
+        ->call('increment')
+        ->call('previousStep')
+        ->emitEvents()->in($this->wizard);
+
+    $this->firstStep
+        ->call('nextStep')
+        ->emitEvents()->in($this->wizard);
+
+    $allStepState = $this->wizard->jsonContent('allStepState');
+    expect($allStepState['second-step']['allStepNames'])->toBe([
+        'first-step',
+        'second-step',
+        'third-step',
+    ]);
+
+    $currentStepState = $this->wizard->jsonContent('currentStepState');
+    expect($currentStepState['counter'])->toBe(1);
+});
+
+it('has a steps property to render navigation', function () {
+    $this->firstStep
+        ->call('nextStep')
+        ->emitEvents()->in($this->wizard);
+
+    $this->firstStep
+        ->call('nextStep')
+        ->emitEvents()->in($this->wizard);
+
+    Livewire::test(SecondStepComponent::class)
+        ->call('increment')
+        ->call('previousStep')
+        ->emitEvents()->in($this->wizard);
+
+    $navigationHtml = $this->wizard->htmlContent('navigation');
+
+    assertMatchesHtmlSnapshot($navigationHtml);
+});
+