+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to `laravel-livewire-wizard` will be documented in this file.
 
+## 0.0.3 - 2022-05-16
+
+## What's Changed
+
+- info function should be stepInfo by @xHeinrich in https://github.com/spatie/laravel-livewire-wizard/pull/1
+- feat: skip step functionality by @ju5t in https://github.com/spatie/laravel-livewire-wizard/pull/3
+- tests: add complete tests for shouldSkip by @ju5t in https://github.com/spatie/laravel-livewire-wizard/pull/5
+
+## New Contributors
+
+- @xHeinrich made their first contribution in https://github.com/spatie/laravel-livewire-wizard/pull/1
+- @ju5t made their first contribution in https://github.com/spatie/laravel-livewire-wizard/pull/3
+
+**Full Changelog**: https://github.com/spatie/laravel-livewire-wizard/compare/0.0.2...0.0.3
+
 ## 0.0.2 - 2022-05-15
 
 **Full Changelog**: https://github.com/spatie/laravel-livewire-wizard/compare/0.0.1...0.0.2